+++ apps/web/src/components/index.ts
@@ -1,4 +1,7 @@
 import "mapbox-gl/dist/mapbox-gl.css";
 
 export { MapPlaceholder } from "./MapPlaceholder";
+export { MapView } from "./MapView";
+export { ColorLegend } from "./ColorLegend";
+export { DiscoveryClient } from "./DiscoveryClient";
 export { PhaseOneHero } from "./PhaseOneHero";