+++ packages/data/src/index.ts
@@ -1,5 +1,7 @@
 export * from "./types";
 export * from "./opportunity-index";
 export * from "./opportunity-index-examples";
+export * from "./metric-utils";
+export * from "./boundaries";
 export * from "./loaders";
 export * from "./validation";